ml_project/requirements.txt
alembic==1.8.1
appnope==0.1.3
astroid==2.12.12
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
cycler==0.11.0
databricks-cli==0.17.3
debugpy==1.6.3
decorator==5.1.1
dill==0.3.6
docker==6.0.0
entrypoints==0.4
Flask==2.2.2
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
greenlet==2.0.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.2
ipython==7.34.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
kiwisolver==1.4.4
lazy-object-proxy==1.8.0
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-dataclass==8.5.9
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mlflow==1.30.0
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.21.6
oauthlib==3.2.2
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.2
prometheus-client==0.15.0
prometheus-flask-exporter==0.20.3
prompt-toolkit==3.0.31
protobuf==4.21.9
psutil==5.9.3
ptyprocess==0.7.0
pyaml==21.10.1
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.15.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
querystring-parser==1.2.4
requests==2.28.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.42
sqlparse==0.4.3
-e git+ssh://git@github.com/made-mlops-2022/dmitrii_korzh.git@fba8b0d4871a1531b8039930e6711f4a301164fe#egg=src&subdirectory=ml_project
tabulate==0.9.0
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.5.0
typed-ast==1.5.4
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
websocket-client==1.4.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
